refactor: Moved utils into data_tools
- Downgraded to using web-optimized spacy language model
