+++ data_tools/__init__.py
@@ -1,3 +1,6 @@
 from .data_separator import *
 from .featurizer import *
 from .descriptor import *
+from .data_loader import *
+from .nlp_tools import *
+from .perf_utils import *
